requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2
django-cors-headers==3.14.0
django-phonenumber-field==7.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-nested-routers==0.93.4
exceptiongroup==1.1.1
filelock==3.12.0
identify==2.5.24
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
phonenumbers==8.13.11
platformdirs==3.5.1
pluggy==1.0.0
pre-commit==3.3.1
pycparser==2.21
PyJWT==2.6.0
pytest==7.3.1
pytest-django==4.5.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.29.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.23.0
